pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.0"
description = "Yield estimation model based on NDVI"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
geojson = "^2.5.0"
rasterio = "^1.1.5"
sat-search = "^0.3.0"
lambda-proxy = "^5.2.1"
intake-stac = "^0.3.0"
boto3 = "^1.16.0"
rio-tiler = "2.0.0rc1"
autopep8 = "^1.5.4"

[tool.poetry.dev-dependencies]
pytest = "*"
sphinx_rtd_theme = "*"
sphinx = "*"
pre_commit = "*"
autopep8 = "^1.5.4"
black = { version = "*", allows-prereleases = true }

[tool.poetry.scripts]
src = "src.cli:cli"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_third_party = ["click"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
